Лаб 3. Работа с БД

+++ GiftShop/GiftShopView/FormCreateProcedure.cs
@@ -112,7 +112,7 @@ namespace GiftShopView
                     CustomerId = Convert.ToInt32(comboBoxCustomer.SelectedValue),
                     SetId = Convert.ToInt32(comboBoxSet.SelectedValue),
                     Count = Convert.ToInt32(textBoxCount.Text),
-                    Sum = Convert.ToInt32(textBoxSum.Text)
+                    Sum = Convert.ToDecimal(textBoxSum.Text)
                 });
                 MessageBox.Show("Сохранение прошло успешно", "Сообщение", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 DialogResult = DialogResult.OK;
